refactor: Move code structure
