+++ base/src/consensus_state.rs
@@ -9,9 +9,9 @@ use crate::{
 
 use mina_crypto::{base58::Base58Encodable, hash::*, signature::PublicKey};
 
-use mina_crypto::hash::{Hashable, VrfOutputHash};
 use derive_more::From;
-use mina_crypto::{hash::*, prelude::*, signature::*};
+use mina_crypto::hash::{Hashable, VrfOutputHash};
+use mina_crypto::prelude::*;
 use serde::{Deserialize, Serialize};
 use wire_type::WireType;
