+++ consensus/Cargo.toml
@@ -28,4 +28,3 @@ proptest = "1.0.0"
 rand = "0.8.4"
 base64 = "0.13"
 time = {version = "0.3", features = ["macros"]}
-wasm-bindgen-test = "0.3"
